complete addComment()

+++ src/data/products.js
@@ -16,7 +16,6 @@ const exportedMethods = {
   },
 
   async createProduct(userId, features, unitPrice){
-
     if (!userId) throw "userId must be provided";
     if (!features) throw "features must be provided";
     if (!unitPrice) throw "unitPrice must be provided";
@@ -48,13 +47,13 @@ const exportedMethods = {
   },  
   
   async deleteProduct(userId, productId){
-
     if (!userId) throw "userId must be provided";
     if (!productId) throw "productId must be provided";
 
     const productsCollection = await products();
     if (!ObjectId.isValid(userId)) throw "id is not a valid ObjectId";
     if (!ObjectId.isValid(productId)) throw "id is not a valid ObjectId";
+    if (!await usersData.getUserById(ObjectId(userId))) throw 'No user Found';
     if (!await this.getProductById(ObjectId(productId))) throw "No Product Found";
 
     const deletionInfo = await productsCollection.deleteOne({ _id: ObjectId(productId) });
@@ -75,8 +74,43 @@ const exportedMethods = {
     return;
   },
 
-  async addComment(userId, producctId, comment){
-    return;
+  async addComment(userId, productId, comment_info){
+    if (!userId) throw "userId must be provided";
+    if (!productId) throw "productId must be provided";
+    if (!comment_info) throw "comment must be provided";
+    if (!comment_info.content) throw "comment content must be provided";
+    if (!comment_info.rating) throw "comment rating must be provided";
+
+    if (typeof comment_info.content !== 'string') throw "comment content must be string";
+    if (typeof comment_info.rating !== 'number') throw "comment rating must be number";
+    if (comment_info.rating < 1 || comment_info.rating > 5) throw "rating is out of range (1-5)";
+
+    const productsCollection = await products();
+    if (!ObjectId.isValid(userId)) throw "id is not a valid ObjectId";
+    if (!ObjectId.isValid(productId)) throw "id is not a valid ObjectId";
+    if (!await usersData.getUserById(ObjectId(userId))) throw 'No user Found';
+    if (!await this.getProductById(ObjectId(productId))) throw "No Product Found";
+
+    const userInfo = await usersData.getUserById(ObjectId(userId));
+    var productInfo = await this.getProductById(ObjectId(productId));
+
+    var commentAddInfo = {};
+    commentAddInfo._id = userInfo._id.toString();
+    commentAddInfo.username = userInfo.username;
+    commentAddInfo.content = comment_info.content;
+    commentAddInfo.rating = comment_info.rating;
+
+    new_overall_score = (productInfo.overall_score * productInfo.comments.length + comment_info.rating)/(productInfo.comments.length + 1);
+    productInfo.comments.push(commentAddInfo);
+
+    var updateInfo = {
+      comments: productInfo.comments,
+      overall_score: new_overall_score
+    }
+
+    await productsCollection.updateOne({ _id: ObjectId(productId) }, { $set: updateInfo });
+
+    return { commentAdded: true };
   }
   
 }
